+++ front/src/features/form/LoadingBox.tsx
