+++ front/src/features/form/PreviewImageBox.tsx
