+++ front/src/features/form/SpoilerSwitch.tsx
